updates
SVN-Revision: 64489

+++ SOC/2009/fusion/libs/fusion/test/suite1/algorithm/all.cpp
@@ -40,19 +40,19 @@ main()
     }
 
     {
-        boost::fusion::vector<int, short, double> t(1, 2, 3.3);
+        boost::fusion::vector<int, long, double> t(1, 2, 3.3);
         BOOST_TEST((boost::fusion::all(t, boost::lambda::_1 < 4)));
         BOOST_TEST((boost::fusion::all(t, boost::lambda::_1 > 0)));
     }
 
     {
-        boost::fusion::vector<int, short, double, long> t(1, 2, 3.3, 2);
+        boost::fusion::vector<int, long, double, long> t(1, 2, 3.3, 2);
         BOOST_TEST((boost::fusion::all(t, boost::lambda::_1 < 4)));
         BOOST_TEST((boost::fusion::all(t, boost::lambda::_1 > 0)));
     }
 
     {
-        boost::fusion::vector<int, short, double> t(1, 2, 3.3);
+        boost::fusion::vector<int, long, double> t(1, 2, 3.3);
         BOOST_TEST((!boost::fusion::all(t, boost::lambda::_1 == 1)));
         BOOST_TEST((!boost::fusion::all(t, boost::lambda::_1 < 3)));
     }
