+++ SOC/2009/fusion/libs/fusion/test/suite1/algorithm/any.cpp
@@ -35,12 +35,12 @@ int
 main()
 {
     {
-        boost::fusion::vector<int, short, double> t(1, 2, 3.3);
+        boost::fusion::vector<int, long, double> t(1, 2, 3.3);
         BOOST_TEST(boost::fusion::any(t, boost::lambda::_1 == 2));
     }
 
     {
-        boost::fusion::vector<int, short, double> t(1, 2, 3.3);
+        boost::fusion::vector<int, long, double> t(1, 2, 3.3);
         BOOST_TEST(!boost::fusion::any(t, boost::lambda::_1 == 3));
     }
