+++ SOC/2009/fusion/libs/fusion/test/suite1/algorithm/count.cpp
@@ -16,12 +16,12 @@ int
 main()
 {
     {
-        boost::fusion::vector<int, short, double> t(1, 1, 1);
+        boost::fusion::vector<int, long, double> t(1, 1, 1);
         BOOST_TEST(boost::fusion::count(t, 1) == 3);
     }
 
     {
-        boost::fusion::vector<int, short, double> t(1, 2, 3.3);
+        boost::fusion::vector<int, long, double> t(1, 2, 3.3);
         BOOST_TEST(boost::fusion::count(t, 3) == 0);
     }
