+++ SOC/2009/fusion/libs/fusion/test/suite1/algorithm/replace_if.cpp
@@ -37,7 +37,7 @@ main()
 
     {
         char const* s = "Ruby";
-        typedef vector<int, short, double, long, char const*, float> vector_type;
+        typedef vector<int, long, double, long, char const*, float> vector_type;
         //TODO
         vector_type t1(1, 2, 3.3, 4, s, 5.5f);
