+++ SOC/2009/fusion/libs/fusion/test/suite1/sequence/adapt_assoc_class_named.cpp
@@ -29,7 +29,7 @@ namespace ns
     class point
     {
     public:
-    
+
         point() : x(0), y(0) {}
         point(int x, int y) : x(x), y(y) {}
             
@@ -103,7 +103,7 @@ main()
         // conversion from adapted::point to vector
         ns::point basep(5, 3);
         adapted::point p(basep);
-        boost::fusion::vector<int, short> v(p);
+        boost::fusion::vector<int, long> v(p);
         v = p;
     }
 
@@ -111,7 +111,7 @@ main()
         // conversion from adated::point to list
         ns::point basep(5, 3);
         adapted::point p(basep);
-        boost::fusion::list<int, short> l(p);
+        boost::fusion::list<int, long> l(p);
         l = p;
     }
