+++ SOC/2009/fusion/libs/fusion/test/suite1/sequence/adapt_class_named.cpp
@@ -112,7 +112,7 @@ main()
         ns::point basep(5, 3);
         adapted::point p(basep);
 
-        vector<int, short> v(p);
+        vector<int, long> v(p);
         v = p;
     }
 
@@ -121,7 +121,7 @@ main()
         ns::point basep(5, 3);
         adapted::point p(basep);
 
-        list<int, short> l(p);
+        list<int, long> l(p);
         l = p;
     }
